scripts/requirements-python.txt
ijson==3.2.3
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0